+++ include/infMat.hpp
@@ -8,4 +8,4 @@
 namespace views = std::views;
 typedef Array<bool, Dynamic, 1> ArrayXb;
 template <typename Singularity, bool SelfInfluence> // concept constrain
-Eigen::ArrayXXf makeInfluenceMatrix(int m, int n, std::span<const ComputeTask> compTaskVec);
+Eigen::ArrayXXd makeInfluenceMatrix(int m, int n, std::span<const ComputeTask> compTaskVec);
