+++ include/solver/dense_solver.hpp
@@ -4,8 +4,8 @@
 #include <iostream>
 
 struct DenseSolver : ISolver {
-  Eigen::VectorXf solve(const Eigen::MatrixXf &lhs,
-                        const Eigen::VectorXf &rhs, float tol=1e-6, std::size_t maxit = 10) override {
+  Eigen::VectorXd solve(const Eigen::MatrixXd &lhs,
+                        const Eigen::VectorXd &rhs, double tol=1e-6, std::size_t maxit = 10) override {
     std::cout << "Solving...\n";
     return lhs.fullPivLu().solve(rhs);
   }
